Move environment storage/retrieval to core.
