+++ src/flatlandasp/features/environments/api.py
@@ -4,11 +4,11 @@ from flatland.envs.observations import GlobalObsForRailEnv
 from flatland.envs.rail_env import RailEnv
 from flatland.envs.rail_generators import rail_from_grid_transition_map
 
-from flatlandasp.core.flatland.static_maps import multi_passing_siding_map
-from flatlandasp.features.environments import environment_crud
-from flatlandasp.features.environments.schemas.environment_data_schema import (
+from flatlandasp.core.flatland import environment_crud
+from flatlandasp.core.flatland.schemas.environment_data_schema import (
     EnvironmentData,
 )
+from flatlandasp.core.flatland.static_maps import multi_passing_siding_map
 
 router = APIRouter()
